Implement Snapshots dialog's skeleton

+++ pkg/rancher-desktop/components/SnapshotCard.vue
@@ -48,21 +48,53 @@ export default Vue.extend<Computed, Methods, Computed, Props>({
 
   methods: {
     async restore() {
-      await this.$store.dispatch('snapshots/restore', this.snapshot.id);
-      ipcRenderer.send('snapshot', {
-        type:   'restore',
-        result: 'success',
-        name:   this.snapshot.name,
-      });
+      const ok = await this.showConfirmationDialog('restore');
+
+      if (ok) {
+        await this.$store.dispatch('snapshots/restore', this.snapshot.id);
+        ipcRenderer.send('snapshot', {
+          type:   'restore',
+          result: 'success',
+          name:   this.snapshot.name,
+        });
+      }
     },
 
     async remove() {
-      await this.$store.dispatch('snapshots/delete', this.snapshot.id);
-      ipcRenderer.send('snapshot', {
-        type:   'delete',
-        result: 'success',
-        name:   this.snapshot.name,
-      });
+      const ok = await this.showConfirmationDialog('delete');
+
+      if (ok) {
+        await this.$store.dispatch('snapshots/delete', this.snapshot.id);
+        ipcRenderer.send('snapshot', {
+          type:   'delete',
+          result: 'success',
+          name:   this.snapshot.name,
+        });
+      }
+    },
+
+    async showConfirmationDialog(type: 'restore' | 'delete') {
+      const confirm = await ipcRenderer.invoke(
+        'show-snapshots-dialog',
+        {
+          window: {
+            message: '',
+            detail:  '',
+            type:    'question',
+            buttons: [
+              this.t(`snapshots.dialog.${ type }.actions.cancel`),
+              this.t(`snapshots.dialog.${ type }.actions.ok`),
+            ],
+            cancelId: 1,
+          },
+          format: {
+            header: this.t(`snapshots.dialog.${ type }.header`),
+            name:   this.snapshot.name,
+          },
+        },
+      );
+
+      return confirm.response;
     },
   },
 });
